Added popup capabilities.
Now when we detect more than one file is being added with rjs-add-to-define
 we prompt the user for which path to use.

Refactored the js2 goto functions to have similar code.

Added rjs-eightify-list, which uses js2-nodes to ensure a list is
no more than 80 chars long.  We use this in rjs-sort-require-paths.

Added lots of documentation.

Added rjs-reset-project which puts rjs into a default state so a
different project can define different aliases and formatters.